Moved all files to a dedicated pages folder. pages will be the dedicated deployment folder and hold the actual webpage and any html/css/js/wasm files. The rest of the repo will be for markdown or simple rust libs that wrap a library to wasm or for files with js or babashka that will be moved or merged into the pages properly.
Going forward, I will use babashka tasks for build
targets to compile markdown to html or to compile
rust to wasm and move it to the pages folder. This
way I can maintain all of the things that are
required for the website in the same place. If
a project becomes a serious web page project, like
a wasm interpreter with a fully interactive editor
or something, then it can be moved to a separate
repo with it's own target for pages to deploy.
then I can link to it from this site rather than
include it.