Untested, but full script for the compile-html script.
Had a problem with the lsp not starting properly
in the project if the bb.edn is included. Turns
out the issue has to do with how the classpath is
determined when the lsp detects it is in a project.
It runs the command `bb print-deps --format classpath`
which depends on Java. Since I am just using babashka
for the project without expecting Clojure or Java
this is a problem. The lsp crashes because the
above command fails. Removing the bb.edn fixes the
problem, but without it there are no tasks and no
reason to use babashka. The other option is to
adjust the settings for the project with the lsp
so that it does not call this command. I chose the
second options, and this adds a config file for
the clojure lsp that sets the specs that trigger
a classpath check to an empty list. For now this is
fine. If the project were to include clojure or
one of the other specs this file would probably
need to be modified, but then it will include
Java and there will be no issue.