Setup a project skeleton of sorts.
Set the main page to a homepage with nothing on
it for now. Set the project page as a portfolio
page to just hold projects. I will likely link this
page for resumes and allow people to explore by
going to my main page if they want.

Added a general function script snippet. Put some
stuff in there for dark mode that allows setting
the mode using local storage. This way a users
setting will persist accross pages and possibly
visits. The current setup is poor as it uses a
snippet to put all the functions in each page.
I think what I want is a script for these that just
needs a script tag to import the file. Then I can
just throw that into the same meta snippet as the
sittle download and have it available for each page
without pasting lots of text, though if the functions
file gets really big and there is a lot of clsj use
I might be better off actually using cljs instead
of scittle.