Creating a gh-pages branch to hold the actual page elements.
The master branch can be the only other branch. The docs folder can be
added to .gitignore and be a build arifact that gets copied (somehow) to
the other branch and commited. Then that branch can be pushed to be deployed.
The actual build process may take a little setting up, but it should be possible
with some shell commands.