Removed all the old files and folders.
Created a simple snippet for adding 2 images underneath a project section title. They will display horizontally on large screens and vertically on small screens. You can click on them to view the full size image in another tab.

Added all of the pages with a clj version.

Updated the build to properly copy all of the resources and to compile the html pages. Had to monkey a bit to allow importing the page content in a variable. It had to be changed to a function.

Changed some minor things, such as adding an id to the page snippet to allow naming the page with a more descriptive name, but still having an id that matches the page name and can be used for things like filtering the main nav so that there is not a link to the page you are on.

Added a new favicon and image for fun.